일정, 프로필, 프로젝트 api 수정

+++ backend/src/main/java/fullcare/backend/schedule/dto/response/ScheduleListResponse.java
@@ -12,9 +12,9 @@ public class ScheduleListResponse {
     //private String content;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private int order;
+    private Long order;
     @Builder
-    public ScheduleListResponse(Long scheduleId, String title,LocalDateTime startDate,LocalDateTime endDate, int order) {
+    public ScheduleListResponse(Long scheduleId, String title,LocalDateTime startDate,LocalDateTime endDate, Long order) {
         this.scheduleId = scheduleId;
         this.title = title;
         this.startDate = startDate;
